delete dashboard page
